fix: changes according to PR #11

+++ db/migrations/20220806190612-create-city.js
@@ -1,24 +1,30 @@
 module.exports = {
   async up(queryInterface, Sequelize) {
-    await queryInterface.createTable('city', {
-      id: {
-        allowNull: false,
-        autoIncrement: true,
-        primaryKey: true,
-        type: Sequelize.INTEGER,
-      },
-      cityName: {
-        allowNull: false,
-        type: Sequelize.STRING,
-      },
-      averageGrade: {
-        allowNull: false,
-        type: Sequelize.DOUBLE,
-        default: 0,
-      },
-    });
+    const isCityAlreadyExists = await queryInterface.tableExists('city');
+    if (!isCityAlreadyExists) {
+      await queryInterface.createTable('city', {
+        id: {
+          allowNull: false,
+          autoIncrement: true,
+          primaryKey: true,
+          type: Sequelize.INTEGER,
+        },
+        cityName: {
+          allowNull: false,
+          type: Sequelize.STRING,
+        },
+        averageGrade: {
+          allowNull: false,
+          type: Sequelize.DOUBLE,
+          default: 0,
+        },
+      });
+    }
   },
   async down(queryInterface, Sequelize) {
-    await queryInterface.dropTable('city');
+    const isCityExists = await queryInterface.tableExists('city');
+    if (isCityExists) {
+      await queryInterface.dropTable('city');
+    }
   },
 };
