+++ db/migrations/20220808131148-create-table-comment.js
@@ -1,24 +1,30 @@
 module.exports = {
   async up(queryInterface, Sequelize) {
-    await queryInterface.createTable('comment', {
-      id: {
-        allowNull: false,
-        autoIncrement: true,
-        primaryKey: true,
-        type: Sequelize.INTEGER,
-      },
-      commentText: {
-        allowNull: false,
-        type: Sequelize.STRING,
-      },
-      cityGrade: {
-        allowNull: false,
-        type: Sequelize.INTEGER,
-      },
-    });
+    const isCommentAlreadyExists = await queryInterface.tableExists('comment');
+    if (!isCommentAlreadyExists) {
+      await queryInterface.createTable('comment', {
+        id: {
+          allowNull: false,
+          autoIncrement: true,
+          primaryKey: true,
+          type: Sequelize.INTEGER,
+        },
+        commentText: {
+          allowNull: false,
+          type: Sequelize.STRING,
+        },
+        cityGrade: {
+          allowNull: false,
+          type: Sequelize.INTEGER,
+        },
+      });
+    }
   },
 
   async down(queryInterface, Sequelize) {
-    await queryInterface.dropTable('comment');
+    const isCommentExists = await queryInterface.tableExists('comment');
+    if (isCommentExists) {
+      await queryInterface.dropTable('comment');
+    }
   },
 };
